fix: split approval policy action helpers
